Implement "projects"
Now a "project" needs to be configured in redis in order for it to the
view to be counted.

A project is a "project.<name_of_project>": "domain name" stored on
redis, it will confirm whether the view comes from the configured
project, so someone can't flood another project with fake views (unless
someone tampers with the Headers, but that's outside of the scope  for
this toy.