Remove domain check
Note to self: Referer of a script in a webpage, is not the original page
where the script is being loaded, this means I can't easily (without
using JS) check the original domain that the endpoint is being requested from.

Now there only needs to be a project set up on redis for it to be
allowed as the ID.

This is minimalistic, so I'm fine with it.